requirements.txt
flask==3.1.1
flask-cors==4.0.0
flask-socketio==5.5.1
tensorflow-cpu==2.15.0
pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
h5py==3.10.0
protobuf==3.20.3
gunicorn==20.1.0